Move Angular code around (#238)
* [angular] Move around source code

Moves the Angular renderer code under `renderers` and the examples under `samples`.

* [angular] Link library to demos

We want the Angular renderer to be under `renderers` while the examples should be under `samples/client`. This doesn't work with the Angular CLI which requires the `angular.json` to be at the root.

These changes work around it by symlinking the renderer back into the samples directory.

* [angular] Update CI config